remove one-time guests script
