+++ inspector/package.json
@@ -9,8 +9,7 @@
     "axios": "^0.20.0",
     "delay": "^4.4.0",
     "dotenv": "^8.2.0",
-    "lodash.countby": "^4.6.0",
-    "lodash.sortby": "^4.7.0"
+    "lodash.countby": "^4.6.0"
   },
   "devDependencies": {
     "jest": "^26.4.2",
